+++ NEWS.md
@@ -2,8 +2,9 @@
 News
 ====
 
-Version 0.17.1 - dev
---------------------
+Version 0.17.1b0 - dev
+----------------------
+
 - CHANGE: using [PySide6](https://pypi.org/project/PySide6/) as Qt binding 
   if installed, `PyQt5` is still supported as fallback
 - NEW: tracking feature for DXF entity copies, new properties of `DXFEntity`